% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{mk_tidy_albums}
\alias{mk_tidy_albums}
\title{Tidy an \code{albums} API response into a flat tibble}
\usage{
mk_tidy_albums(data)
}
\description{
Tidy an \code{albums} API response into a flat tibble
}
\keyword{internal}
