% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{mk_tidy_artists}
\alias{mk_tidy_artists}
\title{Tidy an \code{artists} API response into a flat tibble}
\usage{
mk_tidy_artists(data, image_width = 300, image_height = 300)
}
\description{
Tidy an \code{artists} API response into a flat tibble
}
\keyword{internal}
