% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{mk_tidy_songs}
\alias{mk_tidy_songs}
\title{Tidy a \code{songs} API response into a flat tibble}
\usage{
mk_tidy_songs(data)
}
\description{
Tidy a \code{songs} API response into a flat tibble
}
\keyword{internal}
