% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musickitr-package.R
\docType{package}
\name{musickitr-package}
\alias{musickitr}
\alias{musickitr-package}
\title{musickitr: Tidy Access to the Apple Music API}
\description{
Provides tidy, tibble-based access to the Apple Music Catalog API (search, charts, artists, albums, songs) plus a few similarity utilities useful for building music recommendation and matchup tools. Users must bring their own MusicKit developer credentials (Team ID, Key ID, and private key) from an active Apple Developer Program membership.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavidBrocker/musickitr}
  \item Report bugs at \url{https://github.com/DavidBrocker/musickitr/issues}
}

}
\author{
\strong{Maintainer}: David Brocker \email{davidabrocker@gmail.com}

Authors:
\itemize{
  \item David Brocker \email{davidabrocker@gmail.com}
}

}
\keyword{internal}
